corrects level discrepacy when changing fs, issue #4
A new strategy with frequency oversampling and interpolation at the limits of normalization band solves the issue.
The length of filters in the calculation phase (i.e., freq. oversampling in the frequency domain) can be determined in config, and increased from default value in case you find 1/1000 dB level variation between different fs to be unacceptable :-)